requirements.txt
datasets~=2.20.0
flake8~=7.1.0
GitPython~=3.1.43
jsonlines~=4.0.0
openai~=1.35.12
python-dotenv~=1.0.1
pyverilog~=1.3.0
tqdm~=4.66.4
